need reward tracking and logging

+++ src/ui/PlantCard.java
@@ -64,7 +64,7 @@ public class PlantCard extends JPanel {
         return cooldown;
     }
 
-    public int getCurrentCooldown() {
+    public float getCurrentCooldown() {
         return gameState.getCardCooldown(index);
     }
